requirements.txt
streamlit==1.32.1
python-dotenv
